Change method `KEVM.kompile` to top-level function
